requirements.txt
### Badic ###

alembic>=1.7.6
SQLAlchemy>=1.4.31,<2

fastapi<1,>=0.74.1
uvicorn<1,>=0.17.5
python-jose>=3.3.0

### Others ###

# requests<3,>=2.27.1
